Rebalance: filter messages of revoked partitions
As the `RebalanceSpec` test in https://github.com/akka/alpakka-kafka/pull/865 shows, the consumer stage's
buffer continues to emit messages of revoked partitions that will be re-emitted by a differnt consumer.

This adds another `PartitionAssignmentHandler` which will add filtering in the stage so that
messages of partitions that were revoked are not issued anymore.